Added A SearchForUser java class
This is the JavaBean that will be used when a person search for a user.
It should return a list of all users that fit the search criteria. Fore
example, if a users searches for Jon Doe at San Jose state university,
the query should reflect that as well as the variables in the JavaBean.